requirements.txt
streamlit
pandas
plotly
kaleido
Pillow
reportlab
arabic-reshaper
python-bidi
